alterações de adaptação nas classes

+++ src/main/java/PratoPronto/back/controller/FavoritoController.java
@@ -1,7 +1,6 @@
 package PratoPronto.back.controller;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -26,7 +25,7 @@ public class FavoritoController {
     }
 
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable int id) {
+    public void delete(@PathVariable Long id) {
         favoritoService.deletar(id);
     }
 }
