+++ src/main/java/PratoPronto/back/controller/UploadController.java
@@ -1,7 +1,6 @@
 package PratoPronto.back.controller;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -26,7 +25,7 @@ public class UploadController {
     }
 
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable int id) {
+    public void delete(@PathVariable Long id) {
         uploadService.deletar(id);
     }
 }
