+++ src/main/java/PratoPronto/back/repository/IngredienteRepository.java
@@ -1,9 +1,11 @@
 package PratoPronto.back.repository;
 
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
 import PratoPronto.back.model.Ingrediente;
 
 @Repository
-public interface IngredienteRepository extends JpaRepository<Ingrediente, Integer> {
+public interface IngredienteRepository extends JpaRepository<Ingrediente, Long> {
 }
